pyproject.toml
[project]
name = "coagent-web3"
version = "0.1.0"
description = "Coagent for Wb3."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "coagent-python[a2a] @ git+https://github.com/OpenCSGs/coagent.git@2a00e4f2ef9d4d09488f36402a291e12b1753642",
    "grpcio>=1.73.1",
    "protobuf>=6.31.1",
    "google-api-python-client>=2.176.0",
    "hypercorn>=0.17.3",
    "python-telegram-bot>=22.2",
]

[tool.hatch.build.targets.wheel]
packages = ["coagent_web3"]

[tool.hatch.metadata]
allow-direct-references = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"